Use && instead of comma
Clang gives warnings because the first argument to the comma operator
has no side effects and its value is discarded (even though it may
throw an exception, which is the only thing we're interested in here.)
We can avoid this warning by writing noexcept(expr1) && noexcept(expr2)
instead of noexcept(expr1, expr2).  This also has the benefit of more
directly capturing the semantics; saying expr1, expr2 cannot throw is
the same as saying that expr1 and expr2 both cannot throw.